Allow observing upload progress for custom attachments (#2457)
