+++ Sources/StreamChat/ChatClient.swift
@@ -423,7 +423,11 @@ public class ChatClient {
             MessageSender(messageRepository: messageRepository, database: databaseContainer, apiClient: apiClient),
             NewUserQueryUpdater(database: databaseContainer, apiClient: apiClient),
             MessageEditor(messageRepository: messageRepository, database: databaseContainer, apiClient: apiClient),
-            AttachmentQueueUploader(database: databaseContainer, apiClient: apiClient)
+            AttachmentQueueUploader(
+                database: databaseContainer,
+                apiClient: apiClient,
+                attachmentPostProcessor: config.uploadedAttachmentPostProcessor
+            )
         ]
     }
